Add children to processPaint
