+++ package/src/renderer/components/Paint.tsx
@@ -1,15 +1,8 @@
 import type { ReactNode } from "react";
 import React, { useRef, useMemo, forwardRef, useImperativeHandle } from "react";
 
-import type { SkPaint, SkImageFilter } from "../../skia";
-import {
-  isShader,
-  isMaskFilter,
-  isColorFilter,
-  Skia,
-  isImageFilter,
-  isPathEffect,
-} from "../../skia";
+import type { SkPaint } from "../../skia";
+import { Skia } from "../../skia";
 import type { CustomPaintProps, AnimatedProps } from "../processors";
 import { processPaint } from "../processors";
 import { createDeclaration } from "../nodes";
@@ -27,29 +20,7 @@ export const Paint = forwardRef<SkPaint, AnimatedProps<PaintProps>>(
     const onDeclare = useMemo(
       () =>
         createDeclaration<PaintProps>((paintProps, children, ctx) => {
-          processPaint(paint, ctx.opacity, paintProps);
-          children.forEach((child) => {
-            if (isShader(child)) {
-              paint.setShader(child);
-            } else if (isMaskFilter(child)) {
-              paint.setMaskFilter(child);
-            } else if (isColorFilter(child)) {
-              paint.setColorFilter(child);
-            } else if (isPathEffect(child)) {
-              paint.setPathEffect(child);
-            }
-          });
-          const filters = children.filter(isImageFilter);
-          if (filters.length > 0) {
-            paint.setImageFilter(
-              filters
-                .reverse()
-                .reduce<SkImageFilter | null>(
-                  Skia.ImageFilter.MakeCompose,
-                  null
-                )
-            );
-          }
+          processPaint(paint, ctx.opacity, paintProps, children);
           return paint;
         }),
       [paint]
